refactor(brokers): Split off `web`, `fx-sync-channel` auth brokers.
Why this change?
The long term intent is to completely remove the implicit state
machine from within the views. To do so, the implicit state machine
needs to be made explicit. The state machine is unique for each
browser integration. Auth brokers are created for each integration,
and are currently our best mechanism for creating per-integration
state machines.

What are the broker updates?
* `web` is for people who browse directly to the site.
* `fx-sync` - re-purposed to define generic Sync behaviors, does not
  contain any references to channels. Used when someone signs up for
  Sync and verifies in a 2nd browser.
* `fx-sync-channel` is the former `fx-sync` broker, it defines Sync
  behaviors but has channels.

Since each integration will have its own broker, which defines its own
behaviors, we can remove `isDirectAccess` from the relier.
