+++ app/scripts/models/auth_brokers/fx-sync-web-channel.js
@@ -12,12 +12,12 @@ define(function (require, exports, module) {
 
   const _ = require('underscore');
   const Constants = require('lib/constants');
-  const FxSyncAuthenticationBroker = require('./fx-sync');
+  const FxSyncChannelAuthenticationBroker = require('./fx-sync-channel');
   const WebChannel = require('lib/channels/web');
 
-  const proto = FxSyncAuthenticationBroker.prototype;
+  const proto = FxSyncChannelAuthenticationBroker.prototype;
 
-  const FxSyncWebChannelAuthenticationBroker = FxSyncAuthenticationBroker.extend({
+  const FxSyncWebChannelAuthenticationBroker = FxSyncChannelAuthenticationBroker.extend({
     type: 'fx-sync-web-channel',
 
     defaultCapabilities: _.extend({}, proto.defaultCapabilities, {
