+++ app/tests/test_start.js
@@ -75,9 +75,11 @@ function (Translator, Session) {
     '../tests/spec/models/auth_brokers/fx-firstrun-v2',
     '../tests/spec/models/auth_brokers/fx-ios-v1',
     '../tests/spec/models/auth_brokers/fx-sync',
+    '../tests/spec/models/auth_brokers/fx-sync-channel',
     '../tests/spec/models/auth_brokers/index',
     '../tests/spec/models/auth_brokers/oauth',
     '../tests/spec/models/auth_brokers/oauth-redirect',
+    '../tests/spec/models/auth_brokers/web',
     '../tests/spec/models/device',
     '../tests/spec/models/email-resend',
     '../tests/spec/models/flow',
